requirements.txt
crewai>=0.28.8
crewai_tools>=0.1.7
langchain-google-genai
python-dotenv
lxml
